BUILD SUCCESS: Fix all remaining compilation errors (153 → 0)
- Fix Road type ambiguity with using aliases
- Add Strings.Instance, font fields, _animator, XmlActor stubs
- Fix AudioCue namespace conflicts, add missing methods
- Create BokuBot.cs, Popsy.cs, RockSRO.cs stubs for missing actor types
- Fix Color/Texture type ambiguities
- Add missing members to BokuGame, GameThing, UIGridElement, etc.
- Fix float*Matrix with Matrix.Multiply extension
- Total: 46 files changed, 0 errors, 0 warnings

+++ main/Boku/UI2D/UIGridMaterialElement.cs
@@ -350,8 +350,8 @@ namespace Boku.UI2D
             x = (int)(x - strSize.X * scale * 0.5f + 0.5f);
 
             /// Shadow and text colors, with appropriate opacity.
-            Color darkGrey = new Color(10, 10, 10, byteAlpha);
-            Color transWhite = new Color(255, 255, 255, byteAlpha);
+            Color darkGrey = new Color((int)10, (int)10, (int)10, (int)byteAlpha);
+            Color transWhite = new Color((int)255, (int)255, (int)255, (int)byteAlpha);
 
             // Finally, just do it. Note that we can't put the batching any
             // higher up the food chain, because the generic grid doesn't know
